pos progress

+++ Store/src/main/java/com/store/Store/models/Product.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Entity
@@ -14,11 +15,11 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String name;
     private String brand;
-    private double price;
+    private BigDecimal price;
 
     @Column(columnDefinition = "TEXT")
     private String description;
