+++ Store/src/main/java/com/store/Store/repositories/ProductsRepository.java
@@ -3,6 +3,6 @@ package com.store.Store.repositories;
 import com.store.Store.models.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductsRepository extends JpaRepository<Product, Integer> {
+public interface ProductsRepository extends JpaRepository<Product, Long> {
 
 }
